refactor: Move bounds to their own module
