+++ src/bounds.rs
@@ -1,0 +1,180 @@
+use bit_set::BitSet;
+
+use crate::molecule::{Bond, Element, Molecule};
+
+/// Bounding strategies for the search phase of assembly index calcluation.
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
+pub enum Bound {
+    /// Trivial bound of log_2(# remaining bonds/edges).
+    Log,
+    /// Bound using the length of the shortest integer addition chain defined
+    /// using fragment sizes.
+    Int,
+    /// Bound using the length of the shortest vector addition chain defined
+    /// using fragments' number and types of edges.
+    VecSimple,
+    /// Bound using the length of the shortest vector addition chain defined
+    /// using information about the molecule's number of fragments of size 2.
+    VecSmallFrags,
+    /// TODO
+    CoverSort,
+    /// TODO
+    CoverNoSort,
+    /// TODO
+    CliqueBudget,
+}
+
+/// Edge information used in vector addition chain bounds.
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
+struct EdgeType {
+    bond: Bond,
+    ends: (Element, Element),
+}
+
+/// TODO
+pub fn log_bound(fragments: &[BitSet]) -> usize {
+    let mut size = 0;
+    for f in fragments {
+        size += f.len();
+    }
+
+    size - (size as f32).log2().ceil() as usize
+}
+
+/// TODO
+pub fn int_bound(fragments: &[BitSet], m: usize) -> usize {
+    let mut max_s: usize = 0;
+    let mut frag_sizes: Vec<usize> = Vec::new();
+
+    for f in fragments {
+        frag_sizes.push(f.len());
+    }
+
+    let size_sum: usize = frag_sizes.iter().sum();
+
+    // Test for all sizes m of largest removed duplicate
+    for max in 2..m + 1 {
+        let log = (max as f32).log2().ceil();
+        let mut aux_sum: usize = 0;
+
+        for len in &frag_sizes {
+            aux_sum += (len / max) + (len % max != 0) as usize
+        }
+
+        max_s = max_s.max(size_sum - log as usize - aux_sum);
+    }
+
+    max_s
+}
+
+/// TODO
+// Count number of unique edges in a fragment
+// Helper function for vector bounds
+fn unique_edges(fragment: &BitSet, mol: &Molecule) -> Vec<EdgeType> {
+    let g = mol.graph();
+    let mut nodes: Vec<Element> = Vec::new();
+    for v in g.node_weights() {
+        nodes.push(v.element());
+    }
+    let edges: Vec<petgraph::prelude::EdgeIndex> = g.edge_indices().collect();
+    let weights: Vec<Bond> = g.edge_weights().copied().collect();
+
+    // types will hold an element for every unique edge type in fragment
+    let mut types: Vec<EdgeType> = Vec::new();
+    for idx in fragment.iter() {
+        let bond = weights[idx];
+        let e = edges[idx];
+
+        let (e1, e2) = g.edge_endpoints(e).expect("bad");
+        let e1 = nodes[e1.index()];
+        let e2 = nodes[e2.index()];
+        let ends = if e1 < e2 { (e1, e2) } else { (e2, e1) };
+
+        let edge_type = EdgeType { bond, ends };
+
+        if types.contains(&edge_type) {
+            continue;
+        } else {
+            types.push(edge_type);
+        }
+    }
+
+    types
+}
+
+/// TODO
+pub fn vec_simple_bound(fragments: &[BitSet], m: usize, mol: &Molecule) -> usize {
+    // Calculate s (total number of edges)
+    // Calculate z (number of unique edges)
+    let mut s = 0;
+    for f in fragments {
+        s += f.len();
+    }
+
+    let mut union_set = BitSet::new();
+    for f in fragments {
+        union_set.union_with(f);
+    }
+    let z = unique_edges(&union_set, mol).len();
+
+    (s - z) - ((s - z) as f32 / m as f32).ceil() as usize
+}
+
+/// TODO
+pub fn vec_small_frags_bound(fragments: &[BitSet], m: usize, mol: &Molecule) -> usize {
+    let mut size_two_fragments: Vec<BitSet> = Vec::new();
+    let mut large_fragments: Vec<BitSet> = fragments.to_owned();
+    let mut indices_to_remove: Vec<usize> = Vec::new();
+
+    // Find and remove fragments of size 2
+    for (i, frag) in fragments.iter().enumerate() {
+        if frag.len() == 2 {
+            indices_to_remove.push(i);
+        }
+    }
+    for &index in indices_to_remove.iter().rev() {
+        let removed_bitset = large_fragments.remove(index);
+        size_two_fragments.push(removed_bitset);
+    }
+
+    // Compute z = num unique edges of large_fragments NOT also in size_two_fragments
+    let mut fragments_union = BitSet::new();
+    let mut size_two_fragments_union = BitSet::new();
+    for f in fragments {
+        fragments_union.union_with(f);
+    }
+    for f in size_two_fragments.iter() {
+        size_two_fragments_union.union_with(f);
+    }
+    let z = unique_edges(&fragments_union, mol).len()
+        - unique_edges(&size_two_fragments_union, mol).len();
+
+    // Compute s = total number of edges in fragments
+    // Compute sl = total number of edges in large fragments
+    let mut s = 0;
+    let mut sl = 0;
+    for f in fragments {
+        s += f.len();
+    }
+    for f in large_fragments {
+        sl += f.len();
+    }
+
+    // Find number of unique size two fragments
+    let mut size_two_types: Vec<(EdgeType, EdgeType)> = Vec::new();
+    for f in size_two_fragments.iter() {
+        let mut types = unique_edges(f, mol);
+        types.sort();
+        if types.len() == 1 {
+            size_two_types.push((types[0], types[0]));
+        } else {
+            size_two_types.push((types[0], types[1]));
+        }
+    }
+    size_two_types.sort();
+    size_two_types.dedup();
+
+    s - (z + size_two_types.len() + size_two_fragments.len())
+        - ((sl - z) as f32 / m as f32).ceil() as usize
+}
+
